handle exceptions

+++ src/main/java/com/pinsoft/internship/recipeshare/service/RecipeService.java
@@ -3,6 +3,7 @@ package com.pinsoft.internship.recipeshare.service;
 import com.pinsoft.internship.recipeshare.dto.CreateRecipeRequest;
 import com.pinsoft.internship.recipeshare.entity.Category;
 import com.pinsoft.internship.recipeshare.entity.Recipe;
+import com.pinsoft.internship.recipeshare.exceptions.ApiRequestException;
 import com.pinsoft.internship.recipeshare.repository.CategoryRepository;
 import com.pinsoft.internship.recipeshare.repository.RecipeRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -19,19 +20,32 @@ public class RecipeService {
     CategoryService categoryService;
     public Recipe add(CreateRecipeRequest recipeRequest) {
         Recipe recipe = new Recipe();
-        recipe.setName(recipeRequest.getName());
-        recipe.setExplanation(recipeRequest.getExplanation());
-        Category category = categoryService.getById(recipeRequest.getCategoryId()).get();
-        recipe.setCategory(category);
-        return recipeRepository.save(recipe);
+        if(recipeRequest.getName().isEmpty()){
+            throw new ApiRequestException("Name cannot be empty!");
+        }else{
+            recipe.setName(recipeRequest.getName());
+            recipe.setExplanation(recipeRequest.getExplanation());
+            Category category = categoryService.getById(recipeRequest.getCategoryId()).get();
+            recipe.setCategory(category);
+            return recipeRepository.save(recipe);
+        }
+
     }
     public void delete(Long id) {
-        recipeRepository.deleteById(id);
+        if(recipeRepository.findById(id).isEmpty()){
+            throw new ApiRequestException("Recipe id is not found!");
+        }else{
+            recipeRepository.deleteById(id);
+        }
     }
     public List<Recipe> getAll(){
         return recipeRepository.findAll();
     }
     public Optional<Recipe> getById(Long id){
-        return recipeRepository.findById(id);
+        if(recipeRepository.findById(id).isEmpty()){
+            throw new ApiRequestException("The given id is not exist!");
+        }else{
+            return recipeRepository.findById(id);
+        }
     }
 }
